+++ presentation/src/main/java/com/marvelcomics/brito/presentation/main/ui/compose/MarvelScreenDirections.kt
@@ -4,8 +4,8 @@ import com.marvelcomics.brito.presentation.destinations.DetailCharacterDestScree
 import com.marvelcomics.brito.presentation.destinations.HomeComposeDestScreenDestination
 import com.marvelcomics.brito.presentation.destinations.SearchComposeDestScreenDestination
 
-enum class MarvelScreenDirections(val route: String, val progress: Int) {
-    HOME_SCREEN(HomeComposeDestScreenDestination.route, 20),
-    SEARCH_SCREEN(SearchComposeDestScreenDestination.route, 60),
-    DETAIL_SCREEN(DetailCharacterDestScreenDestination.route, 90)
+enum class MarvelScreenDirections(val route: String, val progress: Float) {
+    HOME_SCREEN(HomeComposeDestScreenDestination.route, 20f),
+    SEARCH_SCREEN(SearchComposeDestScreenDestination.route, 60f),
+    DETAIL_SCREEN(DetailCharacterDestScreenDestination.route, 90f)
 }
